File restructuring, moved the interaction.js and crud.js to modules directory
